[codex:orchestration] extract projection policy for attention and next venue
